taimen: Remove deprecated blob dependencies
In Google's own words: "libhidltransport and libhwbinder have been deprecated and the code is in libhidlbase"
With this knowledge we know that these blobs should still work despite not having libhidltransport and libhwbinder
so we can remove the prebuilt blobs that Google added to the vendor which are hacks to make these blobs work.

Commands used to patch all of these libs:
for i in $(grep -Rln "libhidltransport.so"); do patchelf --remove-needed libhidltransport.so $i; done
for i in $(grep -Rln "libhwbinder.so"); do patchelf --remove-needed libhwbinder.so $i; done

Special thanks to AnanJaser1211 for help with the commands.

Signed-off-by: Lunarixus <Nathanb644@hotmail.co.uk>
Signed-off-by: spezi77 <spezi7713@gmx.net>